Update dependencies and refactor code for compatibility
- Updated dependencies in Cargo.toml to their latest versions, including:
  - toml to 1.0
  - reqwest to 0.13
  - rusqlite to 0.38
  - wasmtime to 41
  - rand to 0.10
  - libp2p to 0.56
  - dialoguer to 0.12
  - indicatif to 0.18
  - console to 0.16
  - cron to 0.15
  - tokio-tungstenite to 0.28
  - notify to 8

- Refactored message sending in discord.rs, slack.rs, and browser.rs to use `.into()` for string conversions.
- Updated WhatsApp pairing code generation to use `rand::random_range`.
- Adjusted parameter type in rusqlite query mapping in store.rs to ensure compatibility with the latest rusqlite version.

+++ crates/claw-memory/src/store.rs
@@ -317,7 +317,7 @@ impl MemoryStore {
             Ok(s) => s,
             Err(_) => return vec![],
         };
-        stmt.query_map(rusqlite::params![limit], |row| {
+        stmt.query_map(rusqlite::params![limit as i64], |row| {
             Ok((
                 row.get::<_, String>(0)?,
                 row.get::<_, String>(1)?,
