+++ crates/claw-server/src/hub.rs
@@ -261,7 +261,7 @@ async fn list_skills(
         .map_err(|_| StatusCode::INTERNAL_SERVER_ERROR)?;
 
     let skills: Vec<HubSkill> = stmt
-        .query_map(rusqlite::params![params.limit, params.offset], |row| {
+        .query_map(rusqlite::params![params.limit as i64, params.offset as i64], |row| {
             Ok(row_to_skill(row))
         })
         .map_err(|_| StatusCode::INTERNAL_SERVER_ERROR)?
